SCTP port added.

+++ src/main.h
@@ -5,15 +5,15 @@
 #define TPE_PROTOS_MAIN_H
 
 typedef struct {
-    int port;
+    uint16_t port;
     char * error_file;
     char * listen_address;
     char * management_address;
-    int management_port;
+    uint16_t management_port;
     char * replacement_msg;
     char * filtered_media_types;
     char * origin_server;
-    int origin_port;
+    uint16_t origin_port;
     char *filter_command;
 
 } options;
